requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cpu
mudes==0.1.0
streamlit==1.7
st-annotated-text==1.0.1
